Add comprehensive tests for key provider standards
