+++ pkgs/standards/swarmauri_keyproviders/tests/functional/test_local_provider_functional.py
@@ -1,0 +1,13 @@
+import pytest
+
+from swarmauri_keyproviders import LocalKeyProvider
+
+
+@pytest.mark.acceptance
+@pytest.mark.asyncio
+async def test_random_and_hkdf() -> None:
+    kp = LocalKeyProvider()
+    rnd = await kp.random_bytes(16)
+    assert len(rnd) == 16
+    out = await kp.hkdf(b"ikm", salt=b"s", info=b"i", length=32)
+    assert len(out) == 32
